refactor(Prober): Adjust model field types and caching logic
- Changed the field types of `SampleSize` and `ComboDistribution` in `ChartState` from `int` to `double` to support more precise data representation
- Made `PullTime` in `SongData` and `Status` private fields, and updated cache expiration check logic to use these fields
- Changed `ChartStatus` and `LevelStatus` in `Status` to dictionary types to improve data access efficiency
- Added a shared instance property to `Status` to implement single-day caching

+++ src/Prober/DivingFish/Models/ChartState.cs
@@ -5,7 +5,7 @@ namespace Limekuma.Prober.DivingFish.Models;
 public record ChartState
 {
     [JsonPropertyName("cnt")]
-    public required int SampleSize { get; init; }
+    public required double SampleSize { get; init; }
 
     [JsonPropertyName("diff")]
     public required string Level { get; init; }
@@ -26,5 +26,5 @@ public record ChartState
     public required List<int> RankDistribution { get; init; }
 
     [JsonPropertyName("fc_dist")]
-    public required List<int> ComboDistribution { get; init; }
+    public required List<double> ComboDistribution { get; init; }
 }
